refactor: move ctx_doctor tests into consolidated server.test.ts
Per CONTRIBUTING.md and maintainer feedback, merge regression tests
into the existing consolidated test file instead of creating a new one.
